requirements.txt
giphy-client==1.0.0
python-telegram-bot==13.11
